декомпозировал календарь на мелкие компоненты
